Fixed physical connection test for ppc102, Changed pytest to run based on flags i.e:unit,default,integration. Combined default and comms test as they were already vitually the same. -Elijah.AB
