+++ tests/mock_ppc102_test.py
@@ -4,6 +4,8 @@ from unittest.mock import patch, MagicMock
 # pylint: disable=import-error,no-name-in-module
 from ppc102 import PPC102_Coms
 import time
+import pytest
+pytestmark = pytest.mark.unit
 
 
 class TestPPC102_Coms(unittest.TestCase):
