+++ tests/physical_fw102c_test.py
@@ -4,7 +4,7 @@
 
 
 import pytest
-pytestmark = pytest.mark.skip("Exclude from default pytest runs")
+pytestmark = pytest.mark.integration
 import sys
 import os
 import unittest
